backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
supabase==2.0.3
PyJWT==2.8.0
python-jose==3.3.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
python-dateutil==2.8.2
